requirements.txt
datasets==2.13.1
Flask==2.0.1
Flask_Cors==4.0.0
matplotlib==3.7.2
numpy==1.25.0
Pillow==10.0.0
Requests==2.31.0
torch==2.0.1
torchvision
tqdm==4.65.0
transformers==4.30.2
Unidecode==1.3.6
Werkzeug==2.3.6
